backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
numpy==1.26.4
pandas==2.1.4
scikit-learn==1.3.2
xarray==2023.12.0
aiohttp==3.9.1
python-multipart==0.0.6
python-dotenv==1.0.0
aiofiles==23.2.1
httpx==0.25.2
netcdf4==1.6.5
h5netcdf==1.3.0
scipy==1.11.4
dask==2023.12.1
cftime==1.6.2
bottleneck==1.3.7
pydap==3.4.0
